+++ FA24_PRN221_3W_G1_KoiFarmShop/KoiFarmShop.Repositories/Models/Order.cs
@@ -7,9 +7,9 @@ namespace KoiFarmShop.Repositories.Models;
 
 public partial class Order
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
-    public long? AccountId { get; set; }
+    public int? AccountId { get; set; }
 
     public string CustomerName { get; set; }
